Traipse 'OpenRPG' {090806-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and
go. Traipse also makes it easy for developers to work on code without fear
of sacrifice. 'Grumpy-Goblin' was created as a stablizing branch in an effort
to remove bugs from core code.

Update Summary:
This build introduces Update Manager. Update Manager is a powerful tool in
open beta that allows users more control over their Mercurial updates.