Traipse 'OpenRPG' {090807-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and
go. Traipse also makes it easy for developers to work on code without fear
of sacrifice. 'Grumpy-Goblin' was created as a stablizing branch in an effort
to remove bugs from core code.

Update Summary:
Major improvements in the Server GUI. Server now looks through XML file for
password (thanks, david), lobby name, and port.  Callbacks and other previous
partial implements work.