Traipse 'OpenRPG' {090808-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and
go. Traipse also makes it easy for developers to work on code without fear
of sacrifice. 'Grumpy-Goblin' was created as a stablizing branch in an effort
to remove bugs from core code.

Update Summary:
This version is stable. ServerGUI works with very few problems, Update Manager
is working quite well, and the software has been improved in overall stability
This update cleans print statements, adds the /tmp/tmp.txt placeholder, and
removes one major bug that caused errors to appear with the ServerGUI