Traipse 'OpenRPG' {090827-02}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for developers to work on code without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy' and adds fixes to the code. 'Ornery-Orc''s main goal is to offer more advanced features and enhance the productivity of the user.

Update Summary:
Update Manager is now in version 0.8. While not every button works, users can now browse the different revisions and their different changesets.  The code has been refined some with feature from Core added to it.  A Crash report is now created if the users software crashes.  Update Manager has been moved to the Traipse Suite menu item, and a Debug Console as been added as well.

Please Check the Controls Panel to see the other files changed.