Traipse 'OpenRPG' {090830-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for developers to work on code without fear of sacrifice. 'Grumpy-Goblin' was designed to stablize the software and make it easy for users to install and run.

Update Summary:
Update dirpath_tools to remove the dependency of the OPENRPG_BASE system key.  The new file is compatible with older location.py files and with the new location.py file.