Traipse 'OpenRPG' {100219-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for
developers to work on code without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy' and adds fixes to
the code. 'Ornery-Orc's main goal is to offer more advanced features and enhance the productivity of the user.

Update Summary (Patch-1)
New Features:
New custom welcome message for servers.

Fixes:
Fix to InterParse which caused Unknowns to break
Fix to a URL reference in mplay_server
Fix to metaserver.cache Template and outdated Meta URLs